Add bash script for firstboot with resize root partition
This script will run with the firstboot.service
at the very first boot of the system.

The resize of the root partition will be performed

The eMMC has 3.9 GByte on RevPis equipped with a CM1 or CM3, but those
equipped with a CM3+ may have 7.8, 15.6 or 31.2 GByte.  Original RevPi images
fit the 3.9 GByte eMMC.  If the eMMC is larger, \fBrevpi-factory-reset\fP
resizes the last partition to the maximum available on the eMMC.  Normally
the last partition is the root partition but users are free to create
custom images for the CM3+ with an additional data partition at the end.
The filesystem on the last partition is expected to be ext4 and will be
resized as well.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>